+++ target.h
@@ -24,6 +24,6 @@ public:
     virtual Vector3<T> getNormal(const Vector3<T> &point) const = 0;
 };
 
-typedef ATarget<float> Target;
+typedef ATarget<double> Target;
 
 #endif //RAYTRACER_TARGET_H
